requirements.txt
requests
python-dotenv
mysql-connector-python
streamlit
pandas
plotly